Portfolio level Earnings Yield Chart Included
In the attached code, I have created an additional column which contains
the portfolio level earnings yield across the respective months.
I am not sure if this is a suitable solution. Each row contains an
observation that represents the portfolio level earingsyield which
corresponds to the applicable month.
I will ahve to take this further in an ISSUE